app/requirements.txt
boto3 == 1.24.9
faiss_cpu==1.7.2
fastapi==0.111.0
numpy==1.23.5
pandas==1.5.3
Pillow==10.3.0
pydantic==1.10.9
python-multipart
Requests==2.32.3
torch==2.3.0
torchvision==0.18.0
uvicorn==0.30.1
